Added Environment.HBM2DDL_AUTO = validate property in Hibernate settings

+++ src/main/resources/db/migration/V1__create_all_tables.sql
@@ -38,7 +38,7 @@ CREATE TABLE account (
     /*id             INTEGER NOT NULL default nextval('account_id_seq'),*/
     id             SERIAL NOT NULL,
     account_type   VARCHAR(30),
-    balance        NUMERIC(10, 2),
+    balance        FLOAT,
     create_date    date default CURRENT_DATE,
     employee_id    INTEGER NOT NULL
 );
